[Update] replace old test case file with new test case file
